feat: extract hook + use it on dashboard
